requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
google-generativeai>=0.8.0
torch>=2.0.0
ffmpeg-python>=0.2.0
